requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
langchain==0.3.0
langchain-community==0.3.0
langchain-groq==0.2.0
groq==0.30.0
sentence-transformers==3.1.1
faiss-cpu==1.9.0
python-dotenv==1.0.1
pdfplumber==0.11.4
reportlab==4.2.2